Move out vg pileup
